Split logic-related Rules tests for convenience
